+++ src/python_bioinformagicks/__init__.py
@@ -4,7 +4,7 @@ Top-level package for python bioinformagicks.
 
 __author__ = """Sylvia N. Michki"""
 __email__ = 'sylvia.michki@gmail.com'
-__version__ = '0.3.0'
+__version__ = '0.4.0'
 
 from . import tools as tl
 from . import plotting as pl