+++ tests/test_plotting.py
@@ -5,7 +5,6 @@ import os
 
 import python_bioinformagicks as bim
 
-
 @pytest.fixture
 def sample_adata():
     """
@@ -36,6 +35,7 @@ def test_plot_grouped_proportions(
 
     TODO: implement image-similarity-based testing.
     """
+    import matplotlib.pyplot as plt
 
     fig = bim.pl.plot_grouped_proportions(
         sample_adata,
@@ -45,6 +45,7 @@ def test_plot_grouped_proportions(
 
     save_path = "./tests/data/plot_grouped_proportions.png"
     fig.savefig(save_path, dpi=300)
+    plt.close(fig)
 
     assert (os.path.getsize(save_path) > 100000)
 
@@ -65,6 +66,7 @@ def test_plot_split_embedding(
 
     TODO: implement image-similarity-based testing.
     """
+    import matplotlib.pyplot as plt
 
     fig = bim.pl.plot_split_embedding(
         sample_adata,
@@ -77,5 +79,57 @@ def test_plot_split_embedding(
 
     save_path = "./tests/data/plot_split_embedding.png"
     fig.savefig(save_path, dpi=150)
+    plt.close(fig)
+
+    assert (os.path.getsize(save_path) > 100000)
+
+
+def test_plot_gprofiler_results(
+    genes = [
+        "Acta2", "Tagln", "Myh10", "Actc1", "Ttn",
+        "Stc1", "Col1a1", "Pdgfra", "Hhip", "Obscn"
+    ],
+    sources = ["GO:BP", "GO:MF"],
+    highlight = True, 
+    organism = "mmusculus",
+    ordered = False,
+    n_terms = 5,
+    sort_by = "FDR",
+    title = "muscle_genes",
+    min_FE = 2,
+):
+    """
+
+    Testing bim.pl.plot_profiler_results
+    
+    Will generate 1 figure with 2 plots, because the query
+    goes against two sources.
+
+    TODO: implement image-similarity-based testing.
+    """
+    import matplotlib.pyplot as plt
+
+    ora_df = bim.tl.do_gprofiler_analysis(
+        genes = genes,
+        sources = sources,
+        ordered = ordered,
+        organism = organism,
+        highlight = highlight,
+    )
+
+    if (highlight):
+        ora_df = ora_df[ora_df["highlighted"]]
+
+    fig = bim.pl.plot_gprofiler_results(
+        ora_df,
+        title = title,
+        sort_by = sort_by,
+        n_terms = n_terms,
+        min_FE = min_FE,
+    )
+
+    save_path = "./tests/data/plot_gprofiler_results.png"
+    fig.savefig(save_path, dpi=150)
+    plt.close(fig)
 
     assert (os.path.getsize(save_path) > 100000)