+++ HISTORY.rst
@@ -2,6 +2,12 @@
 History
 =======
 
+v0.4.0 (2025-06-26)
+--------------------
+* Added bim.pl.plot_gprofiler_analysis.
+* Refactored bim.tl.do_gprofiler_analysis to use requests instead of outdated gProfiler package.
+* Changed bim.tl.call_scSNP to allow setting cb_tag as a parameter.
+
 v0.3.0 (2024-12-08)
 --------------------